# web/src/api/openapi.merged.yaml
openapi: 3.1.0
info:
  title: ATRAG API
  description: ATRAG API Documentation
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /collections/{collection_id}/documents/{document_id}/preview:
    get:
      summary: Get document preview
      description: Get a preview of a specific document
      operationId: get_document_preview
      security:
        - BearerAuth: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: string
        - name: document_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/documentPreview'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /collections/{collection_id}/documents/{document_id}/object:
    get:
      summary: Get document object
      description: Get an object from a specific document
      operationId: get_document_object
      security:
        - BearerAuth: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: string
        - name: document_id
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document object
          content:
            application/octet-stream: {}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
...
  /marketplace/collections/{collection_id}:
    get:
      summary: Get MarketplaceCollection details (read-only)
      description: Get details of a subscribed Collection (read-only access)
      parameters:
        - name: collection_id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      responses:
        '200':
          description: MarketplaceCollection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedCollection'
        '401':
          $ref: '#/components/responses/failResponse'
        '403':
          description: Access denied (need subscription)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          $ref: '#/components/responses/failResponse'
  /marketplace/collections/{collection_id}/documents:
    get:
      summary: List documents in MarketplaceCollection (read-only)
      description: Get document list for a subscribed Collection (read-only mode) with pagination, sorting and search capabilities
      parameters:
        - name: collection_id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-based)
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
        - name: sort_by
          in: query
          schema:
            type: string
            enum:
              - name
              - created
              - updated
              - size
              - status
          description: Field to sort by
        - name: sort_order
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
          description: Sort order
        - name: search
          in: query
          schema:
            type: string
          description: Search documents by name
      responses:
        '200':
          description: Document list (read-only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/documentList'
        '401':
          $ref: '#/components/responses/failResponse'
        '403':
          description: Access denied (need subscription)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          $ref: '#/components/responses/failResponse'
  /marketplace/collections/{collection_id}/documents/{document_id}/preview:
    get:
      summary: Preview document in MarketplaceCollection (read-only)
      description: Get document preview for a subscribed Collection (read-only mode)
      parameters:
        - name: collection_id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: document_id
          in: path
          required: true
          description: Document ID
          schema:
            type: string
      responses:
        '200':
          description: Document preview (read-only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/documentPreview'
        '401':
          $ref: '#/components/responses/failResponse'
        '403':
          description: Access denied (need subscription)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          $ref: '#/components/responses/failResponse'
  /marketplace/collections/{collection_id}/documents/{document_id}/object:
    get:
      summary: Get document object from MarketplaceCollection (read-only)
      description: Get an object from a specific document in a subscribed Collection (read-only mode)
      operationId: get_marketplace_document_object
      parameters:
        - name: collection_id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: document_id
          in: path
          required: true
          description: Document ID
          schema:
            type: string
        - name: path
          in: query
          required: true
          description: Object path within the document
          schema:
            type: string
      responses:
        '200':
          description: Document object (read-only)
          content:
            application/octet-stream: {}
        '206':
          description: Partial content (for range requests)
          content:
            application/octet-stream: {}
        '401':
          $ref: '#/components/responses/failResponse'
        '403':
          description: Access denied (need subscription)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '416':
          description: Range not satisfiable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          $ref: '#/components/responses/failResponse'
  /marketplace/collections/{collection_id}/graph:
    get:
      summary: Get knowledge graph for MarketplaceCollection (read-only)
      description: Get knowledge graph data for a subscribed Collection (read-only mode)
      parameters:
        - name: collection_id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: label
          in: query
          description: Node label filter
          schema:
            type: string
            default: '*'
        - name: max_nodes
          in: query
          description: Maximum number of nodes
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 1000
        - name: max_depth
          in: query
          description: Graph traversal depth
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 3
      responses:
        '200':
          description: Knowledge graph data (read-only)
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      type: object
                  edges:
                    type: array
                    items:
                      type: object
                  collection_id:
                    type: string
                  read_only:
                    type: boolean
        '401':
          $ref: '#/components/responses/failResponse'
        '403':
          description: Access denied (need subscription)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          $ref: '#/components/responses/failResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
    CookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
  schemas:
...
    failResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: '400'
        message:
          type: string
          description: Error message
          example: Invalid request
...
    paginatedResponse:
      type: object
      properties:
        total:
          type: integer
          minimum: 0
          description: Total number of items
          example: 100
        page:
          type: integer
          minimum: 1
          description: Current page number
          example: 1
        page_size:
          type: integer
          minimum: 1
          description: Number of items per page
          example: 10
        total_pages:
          type: integer
          minimum: 1
          description: Total number of pages
          example: 10
        has_next:
          type: boolean
          description: Whether there is a next page
          example: true
        has_prev:
          type: boolean
          description: Whether there is a previous page
          example: false
...
    document:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum:
            - UPLOADED
            - EXPIRED
            - PENDING
            - RUNNING
            - COMPLETE
            - FAILED
            - DELETING
            - DELETED
        vector_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        fulltext_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        graph_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        summary_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        vision_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        vector_index_updated:
          type: string
          format: date-time
          description: Vector index last updated time
        fulltext_index_updated:
          type: string
          format: date-time
          description: Fulltext index last updated time
        graph_index_updated:
          type: string
          format: date-time
          description: Graph index last updated time
        summary_index_updated:
          type: string
          format: date-time
          description: Summary index last updated time
        vision_index_updated:
          type: string
          format: date-time
          description: Vision index last updated time
        summary:
          type: string
          description: Summary of the document
        size:
          type: number
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
...
    documentList:
      type: object
      description: A list of documents with pagination
      allOf:
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/document'
        - $ref: '#/components/schemas/paginatedResponse'
...
    visionChunk:
      type: object
      properties:
        id:
          type: string
        asset_id:
          type: string
        text:
          type: string
        metadata:
          type: object
    documentPreview:
      type: object
      properties:
        doc_object_path:
          type: string
          nullable: true
          description: The path to the document object.
        doc_filename:
          type: string
          description: The name of the document.
        converted_pdf_object_path:
          type: string
          nullable: true
          description: The path to the converted PDF object.
        markdown_content:
          type: string
          description: The markdown content of the document.
        chunks:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              text:
                type: string
              metadata:
                type: object
        vision_chunks:
          type: array
          items:
            $ref: '#/components/schemas/visionChunk'
...
    SharedCollectionConfig:
      type: object
      description: Configuration settings for shared collection features
      properties:
        enable_vector:
          type: boolean
          description: Whether vector search is enabled
          default: true
        enable_fulltext:
          type: boolean
          description: Whether fulltext search is enabled
          default: true
        enable_knowledge_graph:
          type: boolean
          description: Whether knowledge graph is enabled
          default: true
        enable_summary:
          type: boolean
          description: Whether summary generation is enabled
          default: false
        enable_vision:
          type: boolean
          description: Whether vision processing is enabled
          default: false
      required:
        - enable_vector
        - enable_fulltext
        - enable_knowledge_graph
        - enable_summary
        - enable_vision
    SharedCollection:
      type: object
      description: Shared Collection information for marketplace users
      properties:
        id:
          type: string
          description: Collection ID
        title:
          type: string
          description: Collection title
        description:
          type: string
          nullable: true
          description: Collection description
        owner_user_id:
          type: string
          description: Original owner user ID
        owner_username:
          type: string
          nullable: true
          description: Original owner username
        subscription_id:
          type: string
          nullable: true
          description: Subscription record ID (has value if subscribed, null if not subscribed)
        gmt_subscribed:
          type: string
          format: date-time
          nullable: true
          description: Subscription time (only has value when subscribed)
        subscription_count:
          type: integer
          default: 0
          description: Total number of subscriptions
        config:
          $ref: '#/components/schemas/SharedCollectionConfig'
          description: Collection configuration settings
      required:
        - id
        - title
        - owner_user_id
        - subscription_count
        - config
...
  responses:
    failResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: '400'
        message:
          type: string
          description: Error message
          example: Invalid request